``build``: Move core to its own package
