+++ .env.example
@@ -1,0 +1,2 @@
+#!/bin/sh
+APP_NAME=
