+++ composer.json
@@ -10,11 +10,11 @@
         }
     ],
     "require": {
-        "php": ">=8.1"
+        "php": ">=8.1",
+        "mirzahilmi/saphpi-core": "^1.0"
     },
     "autoload": {
         "psr-4": {
-            "Saphpi\\": "core/",
             "Saphpi\\Controllers\\": "controllers/",
             "Saphpi\\Rules\\": "rules/",
             "Saphpi\\Models\\": "models/",
