--- core/Contracts/Validation/DataAwareRule.php
@@ -1,6 +1,0 @@
-<?php
-namespace Saphpi\Contracts\Validation;
-
-interface DataAwareRule {
-    public function setData(array $datas): void;
-}
