--- core/Contracts/Validation/ValidationRule.php
@@ -1,8 +1,0 @@
-<?php
-namespace Saphpi\Contracts\Validation;
-
-interface ValidationRule {
-    public function validate(string $attribute, mixed $value): bool;
-
-    public function error(): string;
-}
