--- core/Request.php
@@ -1,36 +1,0 @@
-<?php
-namespace Saphpi;
-
-class Request {
-    public function getPath(): string {
-        $path = $_SERVER['REQUEST_URI'] ?? '/';
-
-        $queryStringPos = strpos($path, '?');
-        if ($queryStringPos === false) {
-            return $path;
-        }
-
-        $path = substr($path, 0, $queryStringPos);
-        return $path;
-    }
-
-    public function getMethod(): string {
-        return strtolower($_SERVER['REQUEST_METHOD']);
-    }
-
-    public function getBody(): array {
-        $body = [];
-
-        if ($this->getMethod() === 'get') {
-            foreach ($_GET as $key => $_) {
-                $body[$key] = filter_input(INPUT_GET, $key, FILTER_SANITIZE_SPECIAL_CHARS);
-            }
-        } elseif ($this->getMethod() === 'post') {
-            foreach ($_POST as $key => $_) {
-                $body[$key] = filter_input(INPUT_POST, $key, FILTER_SANITIZE_SPECIAL_CHARS);
-            }
-        }
-
-        return $body;
-    }
-}
