--- core/Response.php
@@ -1,12 +1,0 @@
-<?php
-namespace Saphpi;
-
-class Response {
-    public function setHttpStatus(int $code): void {
-        http_response_code($code);
-    }
-
-    public function redirect(string $path): void {
-        header("Location: $path");
-    }
-}
